+++ src/cpp/server/include/server/ServerOptions.gen.hpp
@@ -104,7 +104,7 @@ protected:
       "The port that RStudio Server will bind to while listening for incoming connections. If left empty, the port will be automatically determined based on your SSL settings (443 for SSL, 80 for no SSL).")
       ("www-socket",
       value<std::string>(&wwwSocket_)->default_value(""),
-      "The socket that RStudio Server will bind to while listening for incoming connections. If left empty, a socket will not be used but a port instead.")
+      "The socket that RStudio Server will bind to while listening for incoming connections. If left empty, a port will be used.")
       ("www-root-path",
       value<std::string>(&wwwRootPath_)->default_value(kRequestDefaultRootPath),
       "The path prefix added by a proxy to the incoming RStudio URL. This setting is used so RStudio Server knows what path it is being served from. If running RStudio Server behind a path-modifying proxy, this should be changed to match the base RStudio Server URL.")