+++ NEWS.md
@@ -3,6 +3,7 @@
 ### New
 #### RStudio
 - RStudio now supports the inclusion of environment variables when publishing applications to Posit Connect (#13032)
+- The .Rproj.user folder location can now be customized globally both by users and administrators (#15098)
 - RStudio now supports code formatting using the 'styler' R package, as well via other external applications (#2563)
 - Use native file and message dialogs by default on Linux desktop (#14683; Linux Desktop)
 - Added www-socket option to rserver.conf to enable server to listen on a Unix domain socket (#14938; Open-Source Server)